refactor: refactor the committed
Take out some components, some functions into separate folders
