+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@craco/craco": "^6.4.3",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
@@ -12,12 +13,14 @@
     "@types/react-dom": "^17.0.13",
     "@types/styled-components": "^5.1.24",
     "axios": "^0.26.0",
+    "module-alias": "^2.2.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.6",
-    "react-scripts": "5.0.0",
+    "react-scripts": "^4.0.3",
     "redux": "^4.1.2",
     "redux-devtools-extension": "^2.13.9",
+    "redux-saga": "^1.1.3",
     "redux-thunk": "^2.4.1",
     "sass": "^1.49.9",
     "typescript": "^4.6.2",
