+++ src/components/UpdateButton/style.scss
@@ -1,0 +1,14 @@
+.updateButton {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 60px;
+  height: 50px;
+  border-radius: 10px;
+  background-color: #3b8ea6;
+  transition: 0.3s all ease;
+  &:hover {
+    background-color: #1b6b82;
+    transform: scale(0.97);
+  }
+}
