+++ src/index.tsx
@@ -5,6 +5,7 @@ import { Provider } from "react-redux";
 import App from "./components/App";
 
 import store from "./store/store";
+
 import "./index.scss";
 
 ReactDOM.render(
