+++ src/store/actions/actionsType.ts
@@ -1,0 +1,5 @@
+export const GET_TODOS = "GET_TODO";
+export const ADD_TODO = "ADD_TODO";
+export const TOGGLE_TODO = "TOGGLE_TODO";
+export const REMOVE_TODO = "REMOVE_TODO";
+export const UPDATE_TODO = "UPDATE_TODO";
