+++ src/store/store.ts
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware, combineReducers } from "redux";
 import { composeWithDevTools } from "redux-devtools-extension";
 import thunk from "redux-thunk";
 
-import reducer from "./reducer";
+import reducer from "store/reducer";
 
 const rootReducer = combineReducers({
   todos: reducer,
